requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
cycler==0.11.0
Django==3.2.16
Fiona==1.8.22
fonttools==4.38.0
GDAL @ file:///C:/Users/Ali/Documents/Downloads/GeoSpatiol%20Libraries/GDAL-3.0.4-cp37-cp37m-win_amd64.whl
GeoAlchemy2==0.12.5
geopandas==0.10.2
geoserver-rest==2.3.4
greenlet==2.0.1
idna==3.4
importlib-metadata==5.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
munch==2.5.0
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
psycopg2==2.9.5
Pygments==2.13.0
pyparsing==3.0.9
pyproj==3.2.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
seaborn==0.12.1
Shapely==1.8.5.post1
six==1.16.0
SQLAlchemy==1.4.44
sqlalchemy2-stubs==0.0.2a29
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
xmltodict==0.13.0
zipp==3.11.0
